feat: refact WeekdayCoounter.js
