+++ src/utils/date_formmatter.js
@@ -1,0 +1,8 @@
+function formmatter(date) {
+    const year = date.getFullYear();
+    const month = (date.getMonth() + 1).toString().padStart(2, '0');
+    const day = date.getDate().toString().padStart(2, '0');
+    return `${year}/${month}/${day}`;
+};
+
+export { formmatter as format };
